This script plays the part of updating the list
This script updates the list if it does not exist at the moment, but in the future, might run on a timer considering the node has to run at all times to perform as intended. Will add a search method that uses algorithmically efficient means to check for locally added entries to the list and not just write over them as it currently does.